refactor: split bot into reusable modules
